+++ apps/bridge-dapp/src/pages/Hubble/Wrapper/Wrap/private/useWrapFeeDetailsProps.tsx
@@ -6,9 +6,7 @@ import { notificationApi } from '@webb-tools/webb-ui-components';
 import getViemClient from '@webb-tools/web3-api-provider/utils/getViemClient';
 import { CurrencyConfig } from '@webb-tools/dapp-config/currencies/currency-config.interface';
 import { ensureHex } from '@webb-tools/dapp-config';
-import { WebbWeb3Provider } from '@webb-tools/web3-api-provider';
 import { FungibleTokenWrapper__factory } from '@webb-tools/contracts';
-import { ZERO_ADDRESS } from '@webb-tools/utils';
 import { ZERO_BIG_INT } from '@webb-tools/dapp-config';
 import numberToString from '@webb-tools/webb-ui-components/utils/numberToString';
 import { GasStationFill } from '@webb-tools/icons';
@@ -18,6 +16,7 @@ import type {
   FeeItem,
 } from '@webb-tools/webb-ui-components/src/components/FeeDetails/types';
 
+import { getEstimatedGasFeesByChain } from '../../../../../utils';
 import { AMOUNT_KEY, SOURCE_CHAIN_KEY } from '../../../../../constants';
 
 export default function useWrapFeeDetailsProps({
@@ -37,7 +36,7 @@ export default function useWrapFeeDetailsProps({
     number | undefined
   >();
 
-  const { activeApi, apiConfig } = useWebContext();
+  const { apiConfig } = useWebContext();
 
   const [query] = useQueryParams({
     [AMOUNT_KEY]: StringParam,
@@ -199,74 +198,25 @@ export default function useWrapFeeDetailsProps({
     [client, isValidAmount, srcTypedId, fungibleCfg, amount, srcChainCfg]
   );
 
-  useEffect(
-    () => {
-      const updateGasFees = async () => {
-        if (
-          !client ||
-          !srcTypedId ||
-          !fungibleCfg ||
-          !wrappableCfg ||
-          !activeApi ||
-          !(activeApi instanceof WebbWeb3Provider) ||
-          typeof amount !== 'string' ||
-          amount.length === 0 ||
-          !isValidAmount
-        ) {
-          return;
-        }
+  useEffect(() => {
+    const updateGasFees = async () => {
+      if (!srcTypedId || BigInt(amount ?? 0) === ZERO_BIG_INT) return;
+      try {
+        setIsLoadingGasFees(true);
+        const estimatedGasFees = await getEstimatedGasFeesByChain(srcTypedId);
+        setGasFees(parseFloat(formatEther(estimatedGasFees)));
+      } catch (error) {
+        notificationApi.addToQueue({
+          variant: 'error',
+          message: 'Failed to estimate gas fees',
+        });
+      } finally {
+        setIsLoadingGasFees(false);
+      }
+    };
 
-        const fungibleContractAddr = fungibleCfg.addresses.get(srcTypedId);
-        const wrappableTokenAddr = wrappableCfg.addresses.get(srcTypedId);
-
-        if (!fungibleContractAddr || !wrappableTokenAddr) return;
-
-        const walletClient = activeApi.walletClient;
-
-        if (!walletClient || !walletClient.account) return;
-
-        const wrapTokenAddrHex = ensureHex(wrappableTokenAddr);
-        const fungibleContractHex = ensureHex(fungibleContractAddr);
-
-        try {
-          setIsLoadingGasFees(true);
-
-          const estimatedGas = await client.estimateContractGas({
-            address: fungibleContractHex,
-            abi: FungibleTokenWrapper__factory.abi,
-            functionName: 'wrap',
-            args: [
-              wrapTokenAddrHex,
-              // if native token, amount is 0
-              wrapTokenAddrHex === ZERO_ADDRESS
-                ? parseEther('0')
-                : BigInt(amount),
-            ],
-            account: walletClient.account,
-            // if native token, tx value is equal amount
-            value:
-              wrapTokenAddrHex === ZERO_ADDRESS
-                ? BigInt(amount)
-                : parseEther('0'),
-          });
-
-          const gasPrice = await client.getGasPrice();
-          setGasFees(parseFloat(formatEther(gasPrice)) * Number(estimatedGas));
-        } catch (error) {
-          notificationApi.addToQueue({
-            variant: 'error',
-            message: 'Failed to estimate gas fees',
-          });
-        } finally {
-          setIsLoadingGasFees(false);
-        }
-      };
-
-      updateGasFees();
-    },
-    // prettier-ignore
-    [client, srcTypedId, isValidAmount, fungibleCfg, wrappableCfg, activeApi, amount]
-  );
+    updateGasFees();
+  }, [srcTypedId, amount]);
 
   useEffect(() => {
     if (!isValidAmount) {
