+++ apps/bridge-dapp/src/utils/index.ts
@@ -5,6 +5,7 @@ export * from './errors';
 export { default as formatDateTimeByTimestamp } from './formatDateTimeByTimestamp';
 export * from './getCardTitle';
 export { default as getCurrentTimestamp } from './getCurrentTimestamp';
+export { default as getEstimatedGasFeesByChain } from './getEstimatedGasFeesByChain';
 export { default as getNoteSerializations } from './getNoteSerializations';
 export { default as getParam } from './getParam';
 export { default as getRelayerFeePercentage } from './getRelayerFeePercentage';
